[mapleall]Internal commit msg:     [mapleall] CleanCode 5

+++ src/mapleall/maple_be/src/cg/aarch64/aarch64_memlayout.cpp
@@ -101,8 +101,8 @@ uint32 AArch64MemLayout::ComputeStackSpaceRequirementForCall(StmtNode &stmt,  in
     if (ploc.reg0 != 0) {
       continue;  /* passed in register, so no effect on actual area */
     }
-    sizeOfArgsToStkPass = RoundUp(static_cast<uint64>(ploc.memOffset + ploc.memSize),
-                                  static_cast<uint64>(GetPointerSize()));
+    sizeOfArgsToStkPass = static_cast<uint32>(RoundUp(static_cast<uint64>(ploc.memOffset + ploc.memSize),
+        static_cast<uint64>(GetPointerSize())));
   }
   return sizeOfArgsToStkPass;
 }
